% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_chart.R
\name{AreaChart}
\alias{AreaChart}
\title{AreaChart}
\usage{
AreaChart(
  df,
  x_col,
  y_col,
  fill_col,
  group_col,
  chart_title,
  percent_y_axis = TRUE
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{x_col}{A string representing the column name in \code{df} to be used as the x-axis.}

\item{y_col}{A string representing the column name in \code{df} to be used as the y-axis.}

\item{fill_col}{A string representing the column name in \code{df} to define the fill color of the areas.}

\item{group_col}{A string representing the column name in \code{df} used to group the data into series.}

\item{chart_title}{A string for the title of the chart.}

\item{percent_y_axis}{A logical value indicating whether the y-axis should be formatted as percentages.}
}
\value{
A \code{ggplot2} object representing the interactive area chart.
}
\description{
A function to create an interactive area chart using \code{ggplot2} and \code{plotly} with customizable
options for the x-axis, y-axis, fill color, grouping, and chart title.
}
\details{
This function uses \code{ggplot2} to create an area chart with the following features:
\itemize{
\item Data grouped by the \code{group_col} column for separate area series.
\item Area fill colors defined by the \code{fill_col} column.
\item A chart title and labeled axes.
\item A continuous y-axis scaled to display percentages if \code{percent_y_axis} is \code{TRUE}.
\item Conversion of the \code{ggplot2} object into an interactive \code{plotly} chart.
}
}
\examples{
library(ggplot2)
library(plotly)
test_data <- data.frame(
  Date = seq.Date(from = Sys.Date() - 10, by = "day", length.out = 10),
  Value = runif(10),
  FillGroup = rep(c("Red", "Blue"), each = 5),
  Group = rep(c("A", "B"), each = 5)
)

AreaChart(
  df = test_data,
  x_col = "Date",
  y_col = "Value",
  fill_col = "FillGroup",
  group_col = "Group",
  chart_title = "Interactive Area Chart"
)

}
