% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_class_alloc_chart.R
\name{AssetClassAllocationTable}
\alias{AssetClassAllocationTable}
\title{AssetClassAllocationTable}
\usage{
AssetClassAllocationTable(
  df,
  x_col,
  y_col,
  value_col,
  chart_title,
  l_color = "honeydew1",
  h_color = "darkorange3"
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{x_col}{(string) specifying the column name for the x-axis (e.g., fund names).}

\item{y_col}{(string) specifying the column name for the y-axis (e.g., asset classes).}

\item{value_col}{(string) specifying the column name for the values to be displayed and
used for heatmap coloring.}

\item{chart_title}{(string) specifying the title of the chart.}

\item{l_color}{(string) specifying the low-end color for the heatmap gradient
(default: \code{"honeydew1"}).}

\item{h_color}{(string) specifying the high-end color for the heatmap gradient
(default: \code{"darkorange3"}).}
}
\value{
A heatmap as a \code{ggplot2} object.
}
\description{
This function generates a heatmap to provide a quick overview of asset allocation concentrations
across various funds or accounts and asset classes. Higher values are highlighted using the
specified high color (\code{h_color}, default: \code{darkorange3}), while lower values use the low color
(\code{l_color}, default: \code{honeydew1}), illustrating the intensity of allocation for each combination.
}
\examples{
# Example data
test_df <- data.frame(
  fund = rep(c("Fund A", "Fund B", "Fund C"), each = 3),
  asset_class = rep(c("Equity", "Fixed Income", "Cash"), 3),
  allocation = runif(9, 0, 1)
)

AssetClassAllocationTable(
  df = test_df,
  x_col = "fund",
  y_col = "asset_class",
  value_col = "allocation",
  chart_title = "Asset Allocation Heatmap"
)

}
