% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_ggplot.R
\name{BarChart_ggplot}
\alias{BarChart_ggplot}
\title{Bar Chart with Ordering using ggplot2}
\usage{
BarChart_ggplot(
  df,
  x_col,
  y_col,
  color_col,
  bar_type = c("stack", "group"),
  x_label = "",
  y_label = "",
  order_bars = TRUE,
  ascending = FALSE,
  chart_title = NULL,
  show_bar_text = FALSE,
  chart_flip = FALSE,
  axis_text_size = 12,
  num_abv_grph_size = 4,
  txt_space_abv_grh = 0.005
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{x_col}{A string specifying the column in \code{df} that contains the categorical variable for the x-axis.}

\item{y_col}{A string specifying the column in \code{df} that contains numerical values for the y-axis.}

\item{color_col}{A string specifying the column in \code{df} that determines the fill color of the bars.}

\item{bar_type}{A string indicating the type of the bars. Can be "stack" (default) or "group".}

\item{x_label}{A string specifying the label for the x-axis (default: \code{""}).}

\item{y_label}{A string specifying the label for the y-axis (default: \code{""}).}

\item{order_bars}{A logical value indicating whether to order the bars based on the \code{y_col} values (default: \code{TRUE}).}

\item{ascending}{A logical value indicating whether to order the bars in ascending order (\code{TRUE}) or descending order (\code{FALSE}, default).}

\item{chart_title}{A string specifying the title of the chart (default: \code{NULL}).}

\item{show_bar_text}{A logical value indicating whether to display text on top of the bars (default: \code{FALSE}).}

\item{chart_flip}{A logical value indicating whether to flip the chart coordinates (default: \code{FALSE}).}

\item{axis_text_size}{A numeric value specifying the size of the axis text (default: \code{12}).}

\item{num_abv_grph_size}{A numeric value specifying the size of the text above the bars (default: \code{4}).}

\item{txt_space_abv_grh}{A numeric value specifying the space above the bars for the text labels (default: \code{0.005}).}
}
\value{
A ggplot object representing the bar chart.
}
\description{
This function generates a bar chart using ggplot2 with various customizable options, including bar ordering, text labels, and axis labels.
The bars can be ordered by their height (either ascending or descending) and the chart can be flipped if desired.
}
\examples{
library(ggplot2)
test_df <- data.frame(
  category = c("A", "B", "C"),
  values = c(0.3, 0.5, 0.2),
  color = c("red", "green", "blue")
)
BarChart_ggplot(test_df, x_col = "category", y_col = "values",color_col = "color",
chart_title = "Ordered Bar Chart", show_bar_text = TRUE)
}
