% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart_highcharter.R
\name{BarCharter}
\alias{BarCharter}
\title{Generate a Highcharts Bar or Column Chart}
\usage{
BarCharter(
  df,
  x_col,
  y_col,
  color_col,
  bar_type = c("stack", "group"),
  x_label = "",
  y_label = "",
  order_bars = TRUE,
  ascending = FALSE,
  show_bar_text = FALSE,
  chart_title = NULL,
  chart_flip = FALSE,
  axis_text_size = 12,
  num_abv_grph_size = 4,
  txt_space_abv_grh = 0.005
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{x_col}{A string specifying the column name for the x-axis.}

\item{y_col}{A string specifying the column name for the y-axis.}

\item{color_col}{A string specifying the column name for grouping bars by color.}

\item{bar_type}{A character string indicating the type of bars: either \code{"stack"} (default) or \code{"group"}.}

\item{x_label}{A string specifying the label for the x-axis (default is an empty string).}

\item{y_label}{A string specifying the label for the y-axis (default is an empty string).}

\item{order_bars}{Logical, whether to order bars based on their y-values (default: TRUE).}

\item{ascending}{Logical, whether to order bars in ascending order (default: FALSE, meaning descending order).}

\item{show_bar_text}{Logical, whether to display bar values as text on the bars (default: FALSE).}

\item{chart_title}{A string specifying the chart title (default is NULL).}

\item{chart_flip}{Logical, whether to flip the chart (bar chart instead of column chart) (default: FALSE).}

\item{axis_text_size}{Numeric, font size for axis text labels (default: 12).}

\item{num_abv_grph_size}{Numeric, font size for bar text labels if \code{show_bar_text} is TRUE (default: 4).}

\item{txt_space_abv_grh}{Numeric, spacing adjustment for text above bars (default: 0.005).}
}
\value{
A Highcharts object representing the bar or column chart.
}
\description{
This function creates a bar or column chart using the \code{highcharter} package. It supports stacked or grouped bars,
allows ordering of bars, and provides customization options for axis labels, title, and text styling.
}
