% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_panel.R
\name{BarChartPlot}
\alias{BarChartPlot}
\title{BarChartPlot}
\usage{
BarChartPlot(df, x_col, y_col, width = 0.5, color = "#FF8200")
}
\arguments{
\item{df}{A data frame containing the data for the bar chart.}

\item{x_col}{(string) The column in \code{df} used for the x-axis.}

\item{y_col}{(string) The column in \code{df} used for the y-axis values.}

\item{width}{(numeric) Width of the bars (default is 0.5).}

\item{color}{(hex or string) Color of the bars in hex format (default is "#FF8200").}
}
\value{
A \code{plotly} object containing the bar chart.
}
\description{
Generates a bar chart using \code{ggplot2} and returns it as a plotly object for interactive use.
}
