% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_metrics.R
\name{BondMetrics}
\alias{BondMetrics}
\title{Bond Metrics}
\usage{
BondMetrics(Cube)
}
\arguments{
\item{Cube}{A data frame containing bond data. It must include the following columns:
\itemize{
\item \code{Coupon}: The coupon rate of the bond.
\item \code{Price}: The current price of the bond.
\item \code{ModifiedDuration}: The modified duration of the bond.
\item \code{CreditSpreadDuration}: The credit spread duration of the bond.
\item \code{weight}: The weight associated with each bond.
\item \code{Maturity}: The maturity date of the bond.
\item \code{report_date}: The date of the report.
}}
}
\value{
A data frame with the following additional columns:
\itemize{
\item \code{CurrentYield}: The current yield of the bond.
\item \code{Weighted_MD}: The weighted modified duration of the bond.
\item \code{Weighted_CS}: The weighted credit spread duration of the bond.
\item \code{Weighted_CY}: The weighted current yield of the bond.
\item \code{YearsToMaturity}: The number of years until the bond matures.
\item \code{YearSplit}: A categorical variable indicating the maturity range of the bond.
}
}
\description{
The \code{BondMetrics} function calculates and augments bond metrics from a given data frame.
It computes current yield, weighted metrics (Modified Duration, Credit Spread Duration, and Current Yield)
based on specified weights, and categorizes bonds into maturity year ranges.
}
