% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_biz_date.R
\name{.getLastBizDate}
\alias{.getLastBizDate}
\title{Get the Last Business Date}
\usage{
.getLastBizDate(date, calendar)
}
\arguments{
\item{date}{A \code{Date} object representing the target date.}

\item{calendar}{A vector of \code{Date} objects representing valid business days.}
}
\value{
A \code{Date} object corresponding to the last business date before or equal to \code{date}.
}
\description{
This function finds the most recent business date before or equal to a given date
according to a specified business calendar.
}
\examples{
calendar <- as.Date(c("2024-01-02", "2024-01-03", "2024-01-04"))
.getLastBizDate(as.Date("2024-01-05"), calendar) # Returns "2024-01-04"

}
