% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_validator.R
\name{.inputColumnChecker}
\alias{.inputColumnChecker}
\title{.inputColumnChecker}
\usage{
.inputColumnChecker(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Names of the columns to check for.}
}
\value{
Throws an error if columns are not found.
}
\description{
Check for Input Column Names Passed by the User
}
