% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairlytics-package.R
\docType{package}
\name{fairlytics-package}
\alias{fairlytics}
\alias{fairlytics-package}
\title{fairlytics: The package provides a collection of essential functions for performing data cleaning and processing}
\description{
The `fairlytics` package provides a collection of essential functions developed and used by the team. The package includes a variety of functions for data cleaning, manipulation, visualization and data analysis,
}
\author{
\strong{Maintainer}: Menzi Njakazi \email{menzin@fairtree.com}

}
\keyword{internal}
