% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_generator.R
\name{GenerateBusinessDays}
\alias{GenerateBusinessDays}
\title{Generate Business Days Between Two Dates}
\usage{
GenerateBusinessDays(bgn_date, end_date, calendar = "QuantLib/SouthAfrica")
}
\arguments{
\item{bgn_date}{A \code{Date} object or a string in \code{YYYY-MM-DD} format representing the start date.}

\item{end_date}{A \code{Date} object or a string in \code{YYYY-MM-DD} format representing the end date.}

\item{calendar}{A string specifying the business calendar to use. Defaults to \code{"QuantLib/SouthAfrica"}.}
}
\value{
A tibble containing the sequence of business days between \code{bgn_date} and \code{end_date}.
}
\description{
This function generates a sequence of business days between a given start and end date
using a specified business calendar.
}
\details{
The function relies on the \code{bizdays} package to calculate business days,
and it uses the QuantLib implementation for calendar definitions. Ensure that the calendar is
loaded with \code{load_quantlib_calendars} before calling this function.
}
\examples{
# Example usage:
generateBusinessDays("2025-01-01", "2025-01-31", calendar = "QuantLib/SouthAfrica")

}
