% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_returns.R
\name{GenerateCalendarReturn}
\alias{GenerateCalendarReturn}
\title{Create a Calendar Returns Table}
\usage{
GenerateCalendarReturn(df, date_col, return_col)
}
\arguments{
\item{df}{A data frame containing the date and return columns.}

\item{date_col}{A string specifying the name of the date column in \code{df}.}

\item{return_col}{A string specifying the name of the return column in \code{df}.}
}
\value{
A data frame where each row represents a year, with monthly returns
as columns and a YTD return column.
}
\description{
This function calculates calendar returns based on a given dataset, organizing
the returns by year and month, and computing the Year-To-Date (YTD) return.
}
