% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investment_lookthough.R
\name{InvestmentLookThrough}
\alias{InvestmentLookThrough}
\title{InvestmentLookThrough}
\usage{
InvestmentLookThrough(
  fund_holdings,
  underlying_holdings,
  lookup_fund_ticker,
  pos_value_col,
  tickers_col,
  wgt_col
)
}
\arguments{
\item{fund_holdings}{A data frame representing the exposure of the fund holdings.
It must include the \code{zar_position_value} and \code{weight} columns.}

\item{underlying_holdings}{A data frame representing the underlying securities or holdings within the fund.
It must include a column matching the \code{pos_value_col} parameter.}

\item{lookup_fund_ticker}{specifying the ticker of the underlying fund within the \code{fund_holdings} data frame.
For example, \code{"ABC SJ Equity"}.}

\item{pos_value_col}{A string representing the column name for position values in both data frames.
For example, \code{"zar_position_value"}.}

\item{tickers_col}{A string representing the column name for fund tickers in the \code{fund_holdings} data frame.
For example, \code{"instrument_name"}.}

\item{wgt_col}{A string representing the column name for weights in the \code{fund_holdings} data frame.
For example, \code{"weight"}.}
}
\value{
A data frame containing the rescaled look-through data,
with values adjusted based on the overall fund investment.
The output includes the rescaled position values and updated weights for the underlying securities.
}
\description{
The \code{InvestmentLookThrough} function provides a deeper analysis of the underlying fund investments within a portfolio.
Rather than viewing the fund investment as a single entity,
this function breaks down the fund investment to examine the individual securities or holdings it contains.
This approach offers a more granular understanding of the fund's actual exposure to different assets.
The exposure values are then adjusted or scaled to align with the overall value of the fund investment within a portfolio.
}
\examples{
# Example data for overall fund holdings
fund_holdings <- data.frame(
  instrument_name = c("ABC SJ Equity", "XYZ Bond"),
  zar_position_value = c(1000000, 500000),
  weight = c(0.6, 0.4)
)

# Example data for underlying holdings of the fund "ABC SJ Equity"
underlying_holdings <- data.frame(
  security_name = c("Security A", "Security B", "Security C"),
  zar_position_value = c(600000, 300000, 100000)
)

# Performing the investment look-through analysis
result <- InvestmentLookThrough(
  fund_holdings = fund_holdings,
  underlying_holdings = underlying_holdings,
  lookup_fund_ticker = "ABC SJ Equity",
  pos_value_col = "zar_position_value",
  tickers_col = "instrument_name",
  wgt_col = "weight"
)

print(result)
}
