% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_ggplot.R
\name{LineChart}
\alias{LineChart}
\title{LineChart}
\usage{
LineChart(df, x_col, y_col, color_col, group_col, chart_title)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{x_col}{A string representing the column name in \code{df} to be used as the x-axis.}

\item{y_col}{A string representing the column name in \code{df} to be used as the y-axis.}

\item{color_col}{A string representing the column name in \code{df} to define the color of the lines.}

\item{group_col}{A string representing the column name in \code{df} used to group the data into series.}

\item{chart_title}{A string for the title of the chart.}
}
\value{
A \code{plotly} object representing the interactive line chart.
}
\description{
A function to create an interactive line chart using \code{ggplot2} and \code{plotly} with customizable
options for the x-axis, y-axis, color, grouping, and chart title.
}
\details{
This function uses \code{ggplot2} to create a line chart with the following features:
\itemize{
\item Data grouped by the \code{group_col} column for separate line series.
\item Line colors defined by the \code{color_col} column.
\item A chart title and labeled axes.
\item A continuous y-axis scaled to display percentages.
\item Conversion of the \code{ggplot2} object into an interactive \code{plotly} chart.
}
}
\examples{
library(ggplot2)
library(plotly)
data <- data.frame(
  Date = seq.Date(from = Sys.Date() - 10, by = "day", length.out = 10),
  Value = runif(10),
  ColorGroup = rep(c("Red", "Blue"), each = 5),
  Group = rep(c("A", "B"), each = 5)
)

chart <- LineChart(
  df = data,
  x_col = "Date",
  y_col = "Value",
  color_col = "ColorGroup",
  group_col = "Group",
  chart_title = "Interactive Line Chart"
)
chart

}
