% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart_highcharter.R
\name{LineChartr}
\alias{LineChartr}
\title{LineChartr}
\usage{
LineChartr(df, x_col, y_col, color_col, group_col, chart_title)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{color_col}{A string representing the column name in \code{df} to determine the color of the series (optional).}

\item{group_col}{A string representing the column name in \code{df} used to group the data into series.}

\item{chart_title}{A string for the title of the chart.}

\item{x_label}{A string representing the column name in \code{df} to be used as the x-axis.}

\item{y_label}{A string representing the column name in \code{df} to be used as the y-axis.}
}
\value{
A \code{highchart} object representing the line chart.
}
\description{
A function to create a line chart using the \code{highcharter} library with customizable options
for the x-axis, y-axis, and color grouping. The chart includes exporting functionality
and other stylistic customizations.
}
\details{
This function creates a line chart with the following features:
\itemize{
\item Grouping by the \code{group_col} column for different series.
\item Optional customization of series colors using \code{color_col}.
\item A chart title and labeled axes.
\item Export functionality to download the chart with a white background.
\item A disabled range selector for simplicity.
\item Default styling that disables markers for a cleaner look.
}
}
\examples{
library(highcharter)
library(lubridate)
data <- data.frame(
  Date = seq.Date(from = Sys.Date() - 10, by = "day", length.out = 10),
  Value = rnorm(10),
  Group = rep(c("A", "B"), each = 5)
)
chart <- LineChartr(
  df = data,
  x_label = "Date",
  y_label = "Value",
  color_col = NULL,
  group_col = "Group",
  chart_title = "Sample Line Chart"
)
chart

}
