% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot_wrapper.R
\name{LineChartWrapper}
\alias{LineChartWrapper}
\title{LineChartWrapper}
\usage{
LineChartWrapper(
  df,
  x_col,
  y_col,
  color_col,
  group_col,
  chart_title,
  plot_engine = c("ggplot2", "highcharter")
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{x_col}{A string representing the column name in \code{df} to be used as the x-axis.}

\item{y_col}{A string representing the column name in \code{df} to be used as the y-axis.}

\item{color_col}{A string representing the column name in \code{df} to define the color of the lines.}

\item{group_col}{A string representing the column name in \code{df} used to group the data into series.}

\item{chart_title}{A string for the title of the chart.}

\item{plot_engine}{A string indicating the plotting engine to use. Can be either \code{"ggplot2"} or \code{"highcharter"}.
Default is \code{"ggplot2"}.}
}
\value{
A \code{plotly} object (if \code{"ggplot2"} is chosen) or a \code{highchart} object (if \code{"highcharter"} is chosen)
representing the interactive line chart.
}
\description{
A wrapper function to create an interactive line chart using either \code{ggplot2} or \code{highcharter},
with customizable options for the x-axis, y-axis, color, grouping, and chart title.
}
\details{
This function creates an interactive line chart with the following features:
\itemize{
\item Data grouped by the \code{group_col} column for separate line series.
\item Line colors defined by the \code{color_col} column.
\item A chart title and labeled axes.
\item A continuous y-axis scaled to display percentages.
\item Converts the \code{ggplot2} object into an interactive \code{plotly} chart or creates a \code{highcharter} line chart.
}
}
\examples{
library(ggplot2)
library(plotly)
data <- data.frame(
  Date = seq.Date(from = Sys.Date() - 10, by = "day", length.out = 10),
  Value = runif(10),
  ColorGroup = rep(c("Red", "Blue"), each = 5),
  Group = rep(c("A", "B"), each = 5)
)
chart_ggplot2 <- LineChartWrapper(
  df = data,
  x_col = "Date",
  y_col = "Value",
  color_col = "ColorGroup",
  group_col = "Group",
  chart_title = "Interactive Line Chart (ggplot2)",
  plot_engine = "ggplot2"
)
chart_ggplot2

chart_highcharter <- LineChartWrapper(
  df = data,
  x_col = "Date",
  y_col = "Value",
  color_col = "ColorGroup",
  group_col = "Group",
  chart_title = "Interactive Line Chart (Highcharter)",
  plot_engine = "highcharter"
)
chart_highcharter

}
