% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_lookthough.R
\name{LookThroughFundInvestment}
\alias{LookThroughFundInvestment}
\title{Look Through Fund Investment}
\usage{
LookThroughFundInvestment(holding_exposure, lookup_fund_ticker, account_code)
}
\arguments{
\item{holding_exposure}{A data frame representing the exposure of the fund holdings.
It must include the \code{zar_position_value} and \code{weight} columns.}

\item{lookup_fund_ticker}{specifying the ticker of the underlying fund within the \code{holding_exposure} data frame.
For example, \code{"ABC SJ Equity"}.}

\item{account_code}{representing the account code for the fund,
which is used to locate the specific fund data. For example, \code{"XYZELO01"}.}
}
\value{
A data frame containing the rescaled look-through data,
with values adjusted based on the overall fund investment.
The output includes the rescaled position values and updated weights for the underlying securities.
}
\description{
The \code{LookThroughFundInvestment} function provides a deeper analysis of the underlying fund investments within a fund.
Rather than viewing the fund investment as a single entity,
this function breaks down the fund investment to examine the individual securities or holdings it contains.
This approach offers a more granular understanding of the fund's actual exposure to different assets.
The exposure values are then adjusted or scaled to align with the overall value of the fund investment within a portfolio.
}
\examples{
# Analyzing the underlying investments within a specific fund
LookThroughFundInvestment(holding_exposure = Holdings_data,
                                    lookup_fund_ticker = "ABC SJ Equity",
                                    account_code = "XYZELO01")
}
