% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_wgt_bar_chart.R
\name{RelativeWeightBarChart}
\alias{RelativeWeightBarChart}
\title{RelativeWeightBarChart}
\usage{
RelativeWeightBarChart(
  df,
  instrument_col,
  weight_col,
  fund_label_col,
  txt_abv_grph = 0.05,
  instr_txt_size = 12.5,
  axis_txt_size = 12.5,
  num_size = 4,
  percent = TRUE
)
}
\arguments{
\item{df}{A data frame containing the data to be visualized.}

\item{instrument_col}{(string) specifying the column name for the instrument names
(used on the x-axis).}

\item{weight_col}{(string) specifying the column name for the weights (used on the y-axis).}

\item{fund_label_col}{(string) specifying the column name for fund labels (used for
grouping and facetting).}

\item{txt_abv_grph}{(numeric) controlling the distance from the bar graph to the annotated
value (default: \code{0.05}).}

\item{instr_txt_size}{(numeric) controlling the font size of the instrument text (default: \code{12.5}).}

\item{axis_txt_size}{(numeric) controlling the font size of axis text (default: \code{12.5}).}

\item{num_size}{(numeric) controlling the size of the numbers displayed next to the bars
(default: \code{4}).}

\item{percent}{(logical) indicating whether the y-axis should display percentages (\code{TRUE}) or
raw numbers (\code{FALSE}) (default: \code{TRUE}).}
}
\value{
A \code{plotly::ggplotly} object containing the flipped bar graph.
}
\description{
This function generates a flipped bar graph comprising three vertically aligned bar charts.
It is particularly useful for comparing two funds, with the third chart showing the relative
difference between the two. The bars are colored to represent the funds, and values are annotated
on the bars for easier interpretation.
}
\examples{
# Example data
test_df <- data.frame(
  instrument = c("Instrument A", "Instrument B", "Instrument C"),
  weight = c(0.1, -0.05, 0.2),
  fund_label = c("Fund 1", "Fund 1", "Fund 2")
)
RelativeWeightBarChart(test_df, "instrument", "weight", "fund_label", txt_abv_grph = 0.02)

}
