% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_rdcom.R
\name{SendRDComMail}
\alias{SendRDComMail}
\title{SendRDComMail}
\usage{
SendRDComMail(
  receiver,
  mail_cc,
  subject,
  mail_body = NULL,
  mail_attachment = NULL,
  flag = F
)
}
\arguments{
\item{receiver}{A character vector of email addresses for the primary recipients of the email.}

\item{mail_cc}{A character vector of email addresses for CC recipients. Defaults to an empty string if not specified.}

\item{subject}{A character string specifying the subject of the email.}

\item{mail_body}{A character string specifying the body of the email in HTML format. Defaults to \code{NULL}.
If \code{NULL}, the email will only contain the default Outlook signature.}

\item{mail_attachment}{A character vector containing file paths for attachments. Defaults to \code{NULL}, meaning no attachments are added.}

\item{flag}{A logical value indicating whether to flag the email for reply. Defaults to \code{FALSE}.}
}
\value{
None. This function sends an email and does not return any value.
}
\description{
The \code{SendRDComMail} function sends an email using the RDCOMCLIENT's Outlook service.
It allows you to set recipients, CCs, subject, body, and optional attachments.
Additionally, a flag can be set to mark the email for a reply.
The function also includes a 5-second delay after sending each email to avoid rapid successive emails.
}
\details{
\itemize{
\item \strong{Signature Inclusion}: The email includes the default Outlook signature by appending it to the email body.
\item \strong{Attachments}: You can include one or multiple attachments by providing a vector of file paths.
\item \strong{Delay}: A delay of 5 seconds is added after sending the email to prevent excessive calls to the Outlook API.
}
}
