% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_fund_ERA.R
\name{SeparateFundsERA}
\alias{SeparateFundsERA}
\title{SeparateFundsERA}
\usage{
SeparateFundsERA(data_tbl, fund_name)
}
\arguments{
\item{data_tbl}{A data frame or tibble containing financial data. It should include columns:
\code{Fund}, \code{Date}, \code{Stock}, \code{Ticker}, \code{Sector}, \code{currency}, \code{Price}, \code{Quantity},
\verb{Market Value}, \code{Weight}, and \verb{Fx Close}.}

\item{fund_name}{A character vector specifying the names of the funds to include in the analysis.
The function will only include rows corresponding to these fund names.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{Date}: The date of the record.
\item \code{Fund}: The name of the fund.
\item \code{Stock}: The stock name.
\item \code{Ticker}: The stock ticker symbol.
\item \code{Sector}: The sector to which the stock belongs (with empty sectors replaced by "Cash").
\item \code{currency}: The currency of the investment.
\item \code{Price}: The price of the stock.
\item \code{Quantity}: The quantity of the stock held.
\item \verb{Market Value}: The market value of the stock.
\item \code{Weight}: The weight of the stock within the fund.
\item \verb{Fx Close}: The foreign exchange close value.
}
}
\description{
The \code{SeparateFundsERA} function processes a dataset to filter and format data for a specific fund.
It selects relevant columns, replaces missing or empty values in the \code{Sector} column with "Cash",
and arranges the data by \code{Ticker}. The resulting data frame provides key financial information
for the selected fund.
}
