% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft365_auth.R
\name{SetupMicrosoft365Auth}
\alias{SetupMicrosoft365Auth}
\title{SetupMicrosoft365Auth}
\usage{
SetupMicrosoft365Auth(
  tenant = "fairtree.com",
  app_id = "d44a05d5-c6a5-4bbb-82d2-443123722380"
)
}
\arguments{
\item{tenant}{A string specifying the Azure AD tenant name or ID. Default is "fairtree.com".}

\item{app_id}{A string specifying the client ID (application ID) of your registered Azure AD app.}
}
\value{
A Microsoft 365 authentication token object.
}
\description{
The \code{SetupMicrosoft365Auth} function initializes and creates an authentication token for interacting with Microsoft 365 products.
This token is essential when using functionality such as accessing user information within Microsoft 365 and automating .

\strong{Note:} When \code{SetupMicrosoft365Auth} is run for the first time, it prompts for authorization via the Authenticator app.
}
\examples{
# Initialize the Microsoft 365 token
token <- SetupMicrosoft365Auth()
}
