% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_to_trr.R
\name{TRItoTRR}
\alias{TRItoTRR}
\title{TRItoTRR}
\usage{
TRItoTRR(dfTRI)
}
\arguments{
\item{dfTRI}{TRI dataframe in long format (columns: date, Ticker, TR_Index)}
}
\value{
TRR datafarme
}
\description{
The function takes a TRI object and convert the Index values to Returns#'
}
\examples{
TRItoTRR(dfTRI = TRI_df)
}
