% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_panel.R
\name{WindowPanel}
\alias{WindowPanel}
\title{WindowPanel}
\usage{
WindowPanel(panel_layers, title)
}
\arguments{
\item{panel_layers}{(HTML widget) A list of HTML elements representing individual panel sections.}

\item{title}{(string) A title for the window panel, displayed at the top.}
}
\value{
An HTML element containing the draggable, scrollable window panel.
}
\description{
Wraps multiple collapsible sections in a draggable, scrollable panel for displaying interactive content.
}
