% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_panel.R
\name{AccordionPanel}
\alias{AccordionPanel}
\title{AccordionPanel}
\usage{
AccordionPanel(
  df,
  value_col,
  grouping_col,
  chart_title = NULL,
  color = "#FF8200"
)
}
\arguments{
\item{df}{A data frame containing the data to be displayed in the panel.}

\item{value_col}{(string) The column in \code{df} that contains the values to be displayed in the chart.}

\item{grouping_col}{(string) The column that will be used to group the data fer each panel.}

\item{chart_title}{(string) An optional title for the chart, displayed at the top of the panel.}

\item{color}{(hex or string) A color value (in hex format, default is "#FF8200") to define the bar chart's fill color.}
}
\value{
An HTML element containing the collapsible panel with grouped bar charts based on \code{ggplot2}.
}
\description{
Creates a collapsible UI panel for displaying grouped bar charts in an interactive application.
Each panel section is collapsible and can contain any HTML content, making it ideal for organizing
various interactive elements compactly and accessibly.
}
\examples{
library(shiny)

# Example data
mydata <- data.frame(
  group_value = rep(c("Group A", "Group B", "Group C"), each = 3),
  category = rep(c("Category 1", "Category 2", "Category 3"), times = 3),
  amount = c(10, 20, 15, 25, 35, 30, 40, 50, 45)
)
 AccordionPanel(mydata, value_col = "amount", grouping_col = "group_value",
 chart_title = "Collapsible Bar Charts")
}
