% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funds.R
\name{ExtractFunds}
\alias{ExtractFunds}
\title{Extract Funds}
\usage{
ExtractFunds(
  raw_holdings_data,
  report_col,
  universe_info_col,
  include_benchmarks = T
)
}
\arguments{
\item{raw_holdings_data}{A data frame containing the raw holdings data.
It must include the columns specified by \code{report_col} and \code{universe_info_col}.}

\item{report_col}{A string specifying the name of the column that contains fund or report names.}

\item{universe_info_col}{A string specifying the name of the column containing detailed fund descriptions
(e.g., portfolio, benchmark, and differences).}

\item{include_benchmarks}{A boolean flag indicating whether to include benchmark information in the output.
Default is \code{TRUE}. If set to \code{FALSE}, the benchmark column will be excluded from the result.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{report_name}: The name of the fund or report.
\item \code{port}: The portfolio description extracted from the \code{header_report_universes}.
\item \code{bench}: The benchmark description extracted from the \code{header_report_universes} (included only if \code{includeBenchmarks} is \code{TRUE}).
\item \code{diff}: Any additional differences extracted from the \code{header_report_universes}.
}
}
\description{
#' @description
The \code{ExtractFunds} function processes raw holdings data to extract and format fund information.
It creates a data frame with distinct fund names and their descriptions, separating the descriptions
into components such as portfolio, benchmark, and differences. Users can choose whether to include
benchmark information in the output.
}
